backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
alembic
passlib[bcrypt]
email-validator
python-jose[cryptography]
python-multipart
